+++ packages/expo-video/src/index.ts
@@ -2,3 +2,4 @@ import Video from './NativeVideoModule';
 
 export { VideoView, useVideoPlayer } from './VideoView';
 export { Video };
+export { VideoSource } from './VideoView.types';